+++ CHANGES.rst
@@ -12,9 +12,13 @@ Unreleased
   longer counted in statement totals, which could slightly change your total
   results.
 
+- Configuration values which are file paths will now apply tilde-expansion,
+  closing `issue 589`_.
+
 - Be more flexible about the command name displayed by help, fixing
   `issue 600`_. Thanks, Ben Finney.
 
+.. _issue 589: https://bitbucket.org/ned/coveragepy/issues/589/allow-expansion-in-coveragerc
 .. _issue 600: https://bitbucket.org/ned/coveragepy/issues/600/get-program-name-from-command-line-when